* 64-bit platform warning clean up in the source code.

+++ src/buffer.c
@@ -2,43 +2,43 @@
 #include <string.h>
 #include <stdlib.h>
 
-void buffer_openspacehelper(BUFFER *buff, int amount)
+void buffer_openspacehelper(BUFFER *buff, size_t amount)
 {
-    if (amount + buff->pos < 1024)
-        buffer_resize(buff, 1024);
-    else
-        buffer_resize(buff, amount + buff->pos);
+	if (amount + buff->pos < 1024)
+		buffer_resize(buff, 1024);
+	else
+		buffer_resize(buff, amount + buff->pos);
 }
 
 
-void buffer_resize(BUFFER* buff, int size)
+void buffer_resize(BUFFER* buff, size_t size)
 {
     if (size == 0)
     {
-	if (buff->buffer != (char*)&buff->static_buffer) {
-	    free(buff->buffer);
-	    buff->buffer = (char*)&buff->static_buffer;
-	}
-	return;
+		if (buff->buffer != (char*)&buff->static_buffer) {
+			free(buff->buffer);
+			buff->buffer = (char*)&buff->static_buffer;
+		}
+		return;
     }
 
     if (size < 1024) {
-        buff->buffsize = size;
-	buff->pos = buff->pos > size ? size : buff->pos;
-	if (buff->buffer != (char*)&buff->static_buffer) {
-	    free(buff->buffer);
+		buff->buffsize = size;
+		buff->pos = buff->pos > size ? size : buff->pos;
+		if (buff->buffer != (char*)&buff->static_buffer) {
+			free(buff->buffer);
             buff->buffer = (char*)&buff->static_buffer;
-	}
-        return;
+		}
+		return;
     }
 
-    if (buff->buffer == (char*)&buff->static_buffer) {
-	buff->buffer = (char*)malloc(size);
-	memcpy(buff->buffer, &buff->static_buffer, buff->pos);
-	buff->buffsize = size;
-    } else {
-        buff->buffsize = size > buff->buffsize * 2 ? size : buff->buffsize * 2;
-        buff->buffer = (char*)realloc(buff->buffer, buff->buffsize);
+	if (buff->buffer == (char*)&buff->static_buffer) {
+		buff->buffer = (char*)malloc(size);
+		memcpy(buff->buffer, &buff->static_buffer, buff->pos);
+		buff->buffsize = size;
+	} else {
+		buff->buffsize = size > buff->buffsize * 2 ? size : buff->buffsize * 2;
+		buff->buffer = (char*)realloc(buff->buffer, buff->buffsize);
     }
 }
