+++ src/make.c
@@ -508,7 +508,7 @@ int  md5matchescommandline( TARGET *t )
 	    MD5Init( &context );
 
 	    for ( list = vars->value; list; list = list->next )
-		MD5Update( &context, (unsigned char*)list->string, strlen( list->string ) );
+		MD5Update( &context, (unsigned char*)list->string, (unsigned int)strlen( list->string ) );
 
 	    MD5Final( t->rulemd5sum, &context );
 	    t->rulemd5sumclean = (char)hcache_getrulemd5sum( t );
@@ -1264,7 +1264,7 @@ void make0calcmd5sum( TARGET *t, int source )
     MD5Init( &context );
 
     /* add the path of the file to the sum - it is significant because one command can create more than one file */
-    MD5Update( &context, (unsigned char*)t->name, strlen( t->name ) );
+    MD5Update( &context, (unsigned char*)t->name, (unsigned int)strlen( t->name ) );
 
     if( DEBUG_MD5HASH )
 	printf( "\t\t%s\n", t->name );
@@ -1289,7 +1289,7 @@ void make0calcmd5sum( TARGET *t, int source )
 		LIST *list;
 		for ( list = vars->value; list; list = list->next )
 		{
-		    MD5Update( &context, (unsigned char*)list->string, strlen( list->string ) );
+		    MD5Update( &context, (unsigned char*)list->string, (unsigned int)strlen( list->string ) );
 		    if( DEBUG_MD5HASH )
 			printf( "\t\tCOMMANDLINE: %s\n", list->string );
 		}
@@ -1312,7 +1312,7 @@ void make0calcmd5sum( TARGET *t, int source )
     if ( t->includes )
     {
 	const char* includesStr = "#includes";
-	MD5Update( &context, (unsigned char*)includesStr, strlen( includesStr ) );
+	MD5Update( &context, (unsigned char*)includesStr, (unsigned int)strlen( includesStr ) );
 
 	if( DEBUG_MD5HASH )
 	    printf( "\t\t#includes:\n" );
@@ -1335,7 +1335,7 @@ void make0calcmd5sum( TARGET *t, int source )
 	{
 	    if( DEBUG_MD5HASH )
 		printf( "\t\tdepends: %s %s\n", c->target->name, md5tostring( c->target->buildmd5sum ) );
-	    MD5Update( &context, (unsigned char*)c->target->name, strlen( c->target->name ) );
+	    MD5Update( &context, (unsigned char*)c->target->name, (unsigned int)strlen( c->target->name ) );
 	    MD5Update( &context, c->target->buildmd5sum, sizeof( c->target->buildmd5sum ) );
 	}
     }
